+++ 3-tools/universal-debug-api-companion-webextension/src/common/consts/content--start.ts
@@ -1,3 +1,3 @@
 export * from './entry'
-export const LS_ROOT = '🧩UWDT' // !== UWDA
+export const LS_ROOT = '🧩UWDT' // !== UDA
 export const LS_KEY_ENABLED = `${LS_ROOT}.enabled`