+++ 3-tools/universal-debug-api-companion-webextension/src/common/state/origin.ts
@@ -60,6 +60,9 @@ export function is_origin_eligible(origin: string): boolean {
 	if (origin.startsWith('chrome-extension://'))
 		return false
 
+	if (origin.startsWith('file://'))
+		return false // because no access to LS
+
 	if (origin === 'null') // Firefox about:...
 		return false