+++ 3-tools/universal-debug-api-companion-webextension/CHANGELOG.md
@@ -1,6 +1,12 @@
 # CHANGELOG
 **This tool follows [semver](https://semver.org/).**
 
+## PENDING
+* [fix] edition fields
+* [feat] updated Universal Debug API (browser) with new features
+* [feat] now correctly not support local files, with a warning
+* [chore] bumped dependencies
+
 ## v0.0.1
 * initial release to Chrome Store